Change the types of coin selection error fields.
We now use Natural rather than Integer to indicate fields that can never
be negative.

In response to review feedback:

https://github.com/input-output-hk/cardano-coin-selection/pull/50#discussion_r410287017
https://github.com/input-output-hk/cardano-coin-selection/pull/50#discussion_r410287576

+++ src/library/Cardano/CoinSelection/RandomImprove.hs
@@ -233,11 +233,12 @@ payForOutputs options utxo outputsRequested = do
       | amountAvailable < amountRequested =
           ErrUtxoBalanceInsufficient amountAvailable amountRequested
       | utxoCount < outputCount =
-          ErrUtxoNotFragmentedEnough utxoCount outputCount
+          ErrUtxoNotFragmentedEnough
+              (fromIntegral utxoCount) (fromIntegral outputCount)
       | utxoCount <= inputCountMax =
           ErrUtxoFullyDepleted
       | otherwise =
-          ErrMaximumInputCountExceeded inputCountMax
+          ErrMaximumInputCountExceeded (fromIntegral inputCountMax)
     amountAvailable =
         coinMapValue utxo
     amountRequested =
